Move constants to common
