+++ src/legacy/core_plugins/metrics/common/field_types.js
